Alignment and reactivation of DMM and depth encoder optimizations: - new macro NH_3D_DMM for DMM functionality, including several clean-ups. - new macro NH_3D_SDC for SDC functionality, currently only used for covering common SDC and DMM parts. - new macro NH_3D_ENC_DEPTH for encoder optimizations related to depth. - temporary macro TEMP_SDC_CLEANUP for covering several proposed cleanups related to SDC and DMM.

+++ source/App/TAppEncoder/TAppEncCfg.cpp
@@ -1772,7 +1772,7 @@ Bool TAppEncCfg::parseCfg( Int argc, Char* argv[] )
 
   // allocate slice-based dQP values
 #if NH_MV
-  for (Int i = m_layerIdInNuh.size(); i < m_numberOfLayers; i++ )
+  for (Int i = (Int)m_layerIdInNuh.size(); i < m_numberOfLayers; i++ )
   {
     m_layerIdInNuh.push_back( i == 0 ? 0 : m_layerIdInNuh[ i - 1 ] + 1 ); 
   }
